FREEMARKER-188 Move some template seeding code to "freemarker-generator-cli"
