+++ freemarker-generator-cli/src/test/data/template/application.properties
