+++ freemarker-generator-cli/src/test/data/template/nginx/nginx.conf.ftl
